requirements.txt
aiogram==2.21
ffmpeg-python==0.2.0
python-dotenv==0.20.0
python-slugify==6.1.2
requests==2.28.1
